Create get_substitution_rates.py
This program was used to divide the mutations into three groups; single nucleotide polymorphism, double mutation or long region substitution. This program takes a tab delimited variant summary file as input. This file should contain the Chromosome ID, the position of the mutation, the original nucleotide and the altered nucleotide. The output is a txt file that contains the total number of snp's, the number of long region substitutions, the number of double mutations and a substitution matrix. This matrix contains the total number of substitutions per nucleotide change. These changes are described as A > T, A > C etc.