Create all.py
The script was used to link the substitution information to the genes and generates a tab delimited file that can be used to find the sequences of these mutated genes. The script "all.py" needs a sequence file, this is the fasta file of the metagenome, a mutation file, this is a tab delimited variant summary file, and a gene file, this is the gff file created by RastK, as input. The output for "all.py" is a tab delimited file with the following columns: Chromosome id, original amino acid, altered amino acid, synonymous/non-synonymous and the gene id.