Create all_blast.py
The script was used to link the substitution information to the genes and generates a tab delimited file that can be used to find the sequences of these mutated genes. The script ""all\_blast.py" needs a sequence file, this is the fasta file of the metagenome, a mutation file, this is a tab delimited variant summary file, and a gene file, this is a txt file with the Blast2GO output, as input. The output for "all\_blast.py" is a tab delimited file with the following columns: Chromosome id, original amino acid, altered amino acid, synonymous/non-synonymous, the gene name and the gene id.