Create kegg.py
This script takes the output from the second program (all.py/all_blast.py) and uses this to create a fasta file with the protein sequences of the mutated genes. The script needs a sequence file, this is the fasta file of the metagenome, a mutation file, this is the tab delimited file generated by the second program. The output is a fasta file with the protein sequences of the mutated genes and a tab delimited file with the gene id, the total number of synonymous mutations and the total number of non-synonymous mutations for that gene.